pyproject.toml
[tool.poetry]
name = "locify"
version = "0.0.0"
description = "A library for navigating over large codebases."
authors = ["Hoang Tran <descience.thh10@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "locify/**/*" },
]

[tool.poetry.dependencies]
python = "^3.12"
litellm = "^1.50.0"
tree-sitter = "0.21.3"
grep-ast = "0.3.3"
gitpython = "*"


[tool.poetry.group.tests.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
